refactor: extract document property read builder
